+++ zerver/lib/event_schema.py
@@ -1443,9 +1443,9 @@ typing_person_type = DictType(
     ]
 )
 
-equals_private_or_stream = EnumType(
+equals_direct_or_stream = EnumType(
     [
-        "private",
+        "direct",
         "stream",
     ]
 )
@@ -1454,7 +1454,7 @@ typing_start_event = event_dict_type(
     required_keys=[
         ("type", Equals("typing")),
         ("op", Equals("start")),
-        ("message_type", equals_private_or_stream),
+        ("message_type", equals_direct_or_stream),
         ("sender", typing_person_type),
     ],
     optional_keys=[
@@ -1469,7 +1469,7 @@ typing_stop_event = event_dict_type(
     required_keys=[
         ("type", Equals("typing")),
         ("op", Equals("stop")),
-        ("message_type", equals_private_or_stream),
+        ("message_type", equals_direct_or_stream),
         ("sender", typing_person_type),
     ],
     optional_keys=[